requirements.txt
aiofiles==23.1.0
aiogram @ https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip
aiohttp==3.8.5
aiologger==0.7.0
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.2.0
decorator==5.1.1
deep-translator==1.11.4
emoji==2.8.0
fake-useragent==1.2.1
frozenlist==1.4.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hstspreload==2023.1.1
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==2.8
install==1.3.5
jmespath==1.0.1
loguru==0.7.0
magic-filter==1.0.11
multidict==6.0.4
peewee==3.16.3
py==1.11.0
pydantic==1.10.7
pydantic_core==2.4.0
python-dotenv==1.0.0
requests==2.25.1
retry==0.9.2
rfc3986==2.0.0
sniffio==1.3.0
soupsieve==2.4.1
telegram-handler==1.4.4
typing_extensions==4.7.1
urllib3==1.26.16
yarl==1.9.2
